refactor: aave strategies dir structure and improve imports
