+++ packages/dma-contracts/test/fixtures/factories/aave/open-position-types.ts
@@ -1,3 +1,3 @@
-import { AaveOpenArgs, AaveOpenDependencies } from '@dma-library/strategies/aave/open/open'
+import { AaveOpenArgs, AaveOpenDependencies } from '@dma-library/strategies/aave/multiply/open'
 
 export type OpenPositionTypes = [AaveOpenArgs, AaveOpenDependencies]
