+++ packages/dma-contracts/test/fixtures/factories/steth-eth-earn-aave-position.ts
@@ -14,7 +14,7 @@ import { AaveVersion, strategies } from '@dma-library'
 import {
   AaveV2OpenDependencies,
   AaveV3OpenDependencies,
-} from '@dma-library/strategies/aave/open/open'
+} from '@dma-library/strategies/aave/multiply/open'
 import { RiskRatio } from '@domain'
 import BigNumber from 'bignumber.js'
