+++ packages/dma-contracts/test/fixtures/factories/steth-usdc-multiply-aave-position.ts
@@ -11,7 +11,7 @@ import { AaveVersion, strategies } from '@dma-library'
 import {
   AaveV2OpenDependencies,
   AaveV3OpenDependencies,
-} from '@dma-library/strategies/aave/open/open'
+} from '@dma-library/strategies/aave/multiply/open'
 import { RiskRatio } from '@domain'
 import BigNumber from 'bignumber.js'
