+++ packages/dma-contracts/test/fixtures/factories/wsteth-eth-earn-aave-position.ts
@@ -10,11 +10,10 @@ import {
   StrategyDependenciesAaveV3,
 } from '@dma-contracts/test/fixtures/types'
 import { AaveVersion, strategies } from '@dma-library'
-import { aaveV3UniqueContractName } from '@dma-library/protocols/aave'
 import {
   AaveV2OpenDependencies,
   AaveV3OpenDependencies,
-} from '@dma-library/strategies/aave/open/open'
+} from '@dma-library/strategies/aave/multiply/open'
 import { RiskRatio } from '@domain'
 import BigNumber from 'bignumber.js'
 
@@ -138,10 +137,7 @@ export async function wstethEthEarnAavePosition({
   })
 
   let getPosition
-  if (
-    dependencies.protocol.version === AaveVersion.v3 &&
-    aaveV3UniqueContractName in dependencies.addresses
-  ) {
+  if (isV3(dependencies)) {
     const addresses = dependencies.addresses
     getPosition = async () => {
       return await strategies.aave.v3.view(
