+++ packages/dma-library/src/operations/aave/v2/index.ts
@@ -8,7 +8,9 @@ import { AaveV2OpenOperation, open } from './open'
 import { AaveV2OpenDepositBorrowOperation, openDepositAndBorrow } from './open-deposit-and-borrow'
 import { AaveV2PaybackWithdrawOperation, paybackWithdraw } from './payback-withdraw'
 
-export type { AAVEStrategyAddresses } from './addresses'
+export { AAVEStrategyAddresses } from './addresses'
+export { AdjustRiskDownArgs } from './adjust-risk-down'
+export { AdjustRiskUpArgs } from './adjust-risk-up'
 
 export type AaveV2Operations = {
   adjustRiskDown: AaveV2AdjustDownOperation
