+++ packages/dma-library/src/strategies/aave/common/build-deposit-args.ts
@@ -3,7 +3,7 @@ import { ZERO } from '@dma-common/constants'
 import { DepositArgs } from '@dma-library/operations'
 import { AAVEV2StrategyAddresses } from '@dma-library/operations/aave/v2/addresses'
 import { AAVEV3StrategyAddresses } from '@dma-library/operations/aave/v3'
-import { getAaveTokenAddress } from '@dma-library/strategies'
+import { getAaveTokenAddress } from '@dma-library/strategies/aave/common'
 import { AAVETokens, SwapData } from '@dma-library/types'
 import { WithOptionalSwap } from '@dma-library/types/strategy-params'
 import * as SwapUtils from '@dma-library/utils/swap'
