+++ packages/dma-library/src/strategies/aave/common/index.ts
@@ -6,5 +6,6 @@ export { buildDepositArgs } from './build-deposit-args'
 export { buildSimulation } from './build-simulation'
 export { buildSwap } from './build-swap'
 export { buildTransaction } from './build-transaction'
+export { getAaveTokenAddress, getAaveTokenAddresses } from './get-aave-token-addresses'
 export { isV2 } from './is-v2'
 export { isV3 } from './is-v3'
