+++ packages/dma-library/src/strategies/aave/common/is-v2.ts
@@ -1,4 +1,4 @@
-import { AaveVersion } from '@dma-library/strategies'
+import { AaveVersion } from '@dma-library/types/aave'
 
 export function isV2<
   GeneralDeps extends { protocol: { version: AaveVersion } } | { protocolVersion: AaveVersion },
