+++ packages/dma-library/src/strategies/aave/multiply/adjust/adjust-risk-up.ts
@@ -1,36 +1,33 @@
 import { amountToWei } from '@dma-common/utils/common'
-import { getAaveTokenAddress } from '@dma-library/strategies'
-import { buildOperation } from '@dma-library/strategies/aave/adjust/build-operation'
-import { generateTransition } from '@dma-library/strategies/aave/adjust/generate-transition'
-import { simulatePositionTransition } from '@dma-library/strategies/aave/adjust/simulate-position-transition'
-import {
-  AaveAdjustDependencies,
-  ExtendedAaveAdjustArgs,
-} from '@dma-library/strategies/aave/adjust/types'
+import { getAaveTokenAddress } from '@dma-library/strategies/aave/common'
 import { AAVETokens, PositionTransition } from '@dma-library/types'
 import { feeResolver, getSwapDataHelper } from '@dma-library/utils/swap'
 import BigNumber from 'bignumber.js'
 
-export async function adjustRiskDown(
+import { buildOperation } from './build-operation'
+import { generateTransition } from './generate-transition'
+import { simulatePositionTransition } from './simulate-position-transition'
+import { AaveAdjustDependencies, ExtendedAaveAdjustArgs } from './types'
+
+export async function adjustRiskUp(
   args: ExtendedAaveAdjustArgs,
   dependencies: AaveAdjustDependencies,
 ): Promise<PositionTransition> {
-  const isAdjustDown = true
-  const isAdjustUp = !isAdjustDown
+  const isAdjustUp = true
   const fee = feeResolver(args.collateralToken.symbol, args.debtToken.symbol, {
     isIncreasingRisk: isAdjustUp,
     isEarnPosition: args.positionType === 'Earn',
   })
 
   // Get quote swap
-  const estimatedSwapAmount = amountToWei(new BigNumber(1), args.collateralToken.precision)
+  const estimatedSwapAmount = amountToWei(new BigNumber(1), args.debtToken.precision)
   const { swapData: quoteSwapData } = await getSwapDataHelper<
     typeof dependencies.addresses,
     AAVETokens
   >({
     args: {
-      fromToken: args.collateralToken,
-      toToken: args.debtToken,
+      fromToken: args.debtToken,
+      toToken: args.collateralToken,
       slippage: args.slippage,
       fee,
       swapAmountBeforeFees: estimatedSwapAmount,
@@ -42,13 +39,14 @@ export async function adjustRiskDown(
     },
   })
 
-  // SimulateAdjustDown
-  const { simulatedPositionTransition: simulatedAdjustDown } = await simulatePositionTransition(
+  // SimulateAdjustUp
+  const { simulatedPositionTransition: simulatedAdjustUp } = await simulatePositionTransition(
     isAdjustUp,
     quoteSwapData,
     { ...args, fee },
     dependencies,
-    false,
+    true,
+    dependencies.debug,
   )
 
   // Get accurate swap
@@ -57,11 +55,11 @@ export async function adjustRiskDown(
     AAVETokens
   >({
     args: {
-      fromToken: args.collateralToken,
-      toToken: args.debtToken,
+      fromToken: args.debtToken,
+      toToken: args.collateralToken,
       slippage: args.slippage,
       fee,
-      swapAmountBeforeFees: simulatedAdjustDown.swap.fromTokenAmount,
+      swapAmountBeforeFees: simulatedAdjustUp.swap.fromTokenAmount,
     },
     addresses: dependencies.addresses,
     services: {
@@ -74,7 +72,7 @@ export async function adjustRiskDown(
   const operation = await buildOperation({
     adjustRiskUp: isAdjustUp,
     swapData,
-    simulatedPositionTransition: simulatedAdjustDown,
+    simulatedPositionTransition: simulatedAdjustUp,
     collectFeeFrom,
     args,
     dependencies,
@@ -90,7 +88,7 @@ export async function adjustRiskDown(
     operation,
     collectFeeFrom,
     fee,
-    simulatedPositionTransition: simulatedAdjustDown,
+    simulatedPositionTransition: simulatedAdjustUp,
     args,
     dependencies,
     quoteSwapData,
