+++ packages/dma-library/src/strategies/aave/multiply/adjust/adjust.ts
@@ -1,9 +1,9 @@
-import { adjustRiskDown } from '@dma-library/strategies/aave/adjust/adjust-risk-down'
-import { adjustRiskUp } from '@dma-library/strategies/aave/adjust/adjust-risk-up'
 import { getFlashloanToken } from '@dma-library/strategies/aave/common'
 import { PositionTransition } from '@dma-library/types'
 import { isRiskIncreasing } from '@domain/utils/risk-direction'
 
+import { adjustRiskDown } from './adjust-risk-down'
+import { adjustRiskUp } from './adjust-risk-up'
 import { AaveAdjustArgs, AaveAdjustDependencies, ExtendedAaveAdjustArgs } from './types'
 
 export async function adjust(
