+++ packages/dma-library/src/strategies/aave/multiply/adjust/generate-transition.ts
@@ -1,9 +1,10 @@
 import { FEE_ESTIMATE_INFLATOR, ONE, ZERO } from '@dma-common/constants'
 import { amountFromWei } from '@dma-common/utils/common'
 import { calculateFee } from '@dma-common/utils/swap'
-import { GenerateTransitionArgs } from '@dma-library/strategies/aave/adjust/types'
 import { PositionTransition } from '@dma-library/types'
 import BigNumber from 'bignumber.js'
+
+import { GenerateTransitionArgs } from './types'
 
 export async function generateTransition({
   isIncreasingRisk,
