+++ packages/dma-library/src/strategies/aave/multiply/close/build-operation.ts
@@ -5,16 +5,14 @@ import { operations } from '@dma-library/operations'
 import { AAVEStrategyAddresses } from '@dma-library/operations/aave/v2'
 import { AAVEV3StrategyAddresses } from '@dma-library/operations/aave/v3'
 import { CloseArgs } from '@dma-library/operations/aave/v3/close'
-import { AaveVersion } from '@dma-library/strategies'
-import {
-  AaveCloseDependencies,
-  ExpandedAaveCloseArgs,
-} from '@dma-library/strategies/aave/close/types'
 import { IOperation, SwapData } from '@dma-library/types'
+import { AaveVersion } from '@dma-library/types/aave'
 import { resolveFlashloanProvider } from '@dma-library/utils/flashloan/resolve-provider'
 import { feeResolver } from '@dma-library/utils/swap'
 import { FLASHLOAN_SAFETY_MARGIN } from '@domain/constants'
 import BigNumber from 'bignumber.js'
+
+import { AaveCloseDependencies, ExpandedAaveCloseArgs } from './types'
 
 export async function buildOperation(
   swapData: SwapData & {
