+++ packages/dma-library/src/strategies/aave/multiply/open/index.ts
@@ -1,0 +1,9 @@
+import {
+  AaveOpenArgs,
+  AaveOpenDependencies,
+  AaveV2OpenDependencies,
+  AaveV3OpenDependencies,
+} from './open'
+
+export { AaveOpenArgs, AaveOpenDependencies, AaveV2OpenDependencies, AaveV3OpenDependencies }
+export { open } from './open'
