--- packages/dma-library/src/strategies/aave/open/index.ts
@@ -1,1 +1,0 @@
-export { open } from './open'
