+++ packages/dma-library/src/strategies/index.ts
@@ -2,8 +2,6 @@ import { aave } from './aave'
 import { ajna } from './ajna'
 import buckets from './ajna/earn/buckets.json'
 
-export { AaveVersion, getAaveTokenAddress } from './aave'
-
 export const strategies: {
   aave: typeof aave
   ajna: typeof ajna
