+++ packages/dma-library/src/types/aave/index.ts
@@ -1,2 +1,3 @@
 export { AavePosition } from './aave-position'
 export type { AAVETokens } from './tokens'
+export { AaveVersion } from './version'
