+++ packages/dma-library/src/types/aave/version.ts
@@ -1,0 +1,4 @@
+export enum AaveVersion {
+  v2 = 'v2',
+  v3 = 'v3',
+}
